pyproject.toml
[project]
name = "a320-hx-rul"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pyyaml>=6.0.3",
    "torch",
    "tqdm>=4.67.3",
]

[tool.uv.sources]
torch = { path = "wheels/torch-2.11.0+cu130-cp314-cp314-win_amd64.whl" }
